Port the main React components to the new harness
