+++ app/tb/src/components/InfoEd/index.js
