+++ app/tb/src/components/SymptomOverview/Symptoms.css
