+++ app/tb/src/containers/Logout/index.js
