+++ app/tb/src/middleware/loggingMiddleware/index.js
