+++ app/tb/src/util/token.js
