--- app/tbapp/src/containers/App/App.css
@@ -1,44 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-header {
-  background-color: rgba(0, 255, 0, 0.3);
-  height: 150px;
-  padding: 20px;
-  color: black;
-}
-
-.App-title {
-  font-size: 1.5em;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-.App-body {
-  font-size: small;
-}
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
-}
-
-.jumbotron {
-  background-image: url(https://www.mpowercare.org/wp-content/uploads/2016/08/image21.jpg);
-  background-repeat: repeat-x;
-  background-position: center;
-}
-.jumbotron h1 {
-  color: #eee;
-  text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
-
-}
-
-.jumbotron p {
-  color: #eee;
-  text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
-
-}
